Fix asset paths and consolidate point history; add utils and tests
