+++ index.html
@@ -7,7 +7,7 @@
     <link href="https://fonts.googleapis.com/css2?family=Inter:wght@300;400;500;600;700;800&display=swap" rel="stylesheet">
     <link href="https://fonts.googleapis.com/css2?family=Orbitron:wght@400;500;600;700;800;900&display=swap" rel="stylesheet">
     <link href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.4.0/css/all.min.css" rel="stylesheet">
-    <link rel="stylesheet" href="styles.css">
+    <link rel="stylesheet" href="css/styles.css">
 </head>
 <body>
     <nav class="navigation">
@@ -473,7 +473,7 @@
                 </div>
             </div>
             <div class="point-history-section">
-                <h4 style="margin: 15px 0 10px 0; font-size: 14px; font-weight: 600; color: var(--text-primary); font-family: 'Orbitron', monospace;">直近3日のポイント履歴</h4>
+                <h4 style="margin: 15px 0 10px 0; font-size: 14px; font-weight: 600; color: var(--text-primary); font-family: 'Orbitron', monospace;">最近のポイント履歴</h4>
                 <div id="pointHistoryContainer">
                     <!-- Point history will be rendered here -->
                 </div>
@@ -485,6 +485,7 @@
         </div>
     </div>
 
-    <script src="script.js"></script>
+    <script src="js/utils.js"></script>
+    <script src="js/script.js"></script>
 </body>
 </html>
