+++ js/script.js
@@ -1323,7 +1323,7 @@ function completeTaskFromGantt(taskId) {
         // ポイント追加
         const points = task.points || getDefaultPoints(task.priority);
         focusPoints += points;
-        addPointHistory('タスク完了', points);
+        addPointHistory(points, "タスク完了");
         
         saveData();
         updateStats();
@@ -1465,27 +1465,6 @@ function getDefaultPoints(priority) {
     return pointsMap[priority] || 15;
 }
 
-function addPointHistory(action, points) {
-    const today = new Date().toISOString().split('T')[0];
-    
-    if (!pointHistory) {
-        pointHistory = [];
-    }
-    
-    pointHistory.unshift({
-        date: today,
-        action: action,
-        points: points,
-        timestamp: new Date().toISOString()
-    });
-    
-    // 履歴は最新30件まで保持
-    if (pointHistory.length > 30) {
-        pointHistory = pointHistory.slice(0, 30);
-    }
-    
-    console.log('📝 ポイント履歴追加:', action, points);
-}
 
 // CSV Import Functions
 function downloadCSVTemplate() {
@@ -1663,60 +1642,6 @@ function parseCSV(csvText) {
     console.log('✅ CSVパース完了');
 }
 
-function parseCSVLine(line) {
-    const result = [];
-    let current = '';
-    let inQuotes = false;
-    let i = 0;
-    
-    // 行の前後の空白を除去
-    line = line.trim();
-    
-    while (i < line.length) {
-        const char = line[i];
-        
-        if (char === '"') {
-            if (!inQuotes) {
-                // 引用符開始
-                inQuotes = true;
-            } else if (i + 1 < line.length && line[i + 1] === '"') {
-                // エスケープされた引用符 ("")
-                current += '"';
-                i++; // 次の引用符をスキップ
-            } else {
-                // 引用符終了
-                inQuotes = false;
-            }
-        } else if (char === ',' && !inQuotes) {
-            // フィールド区切り
-            result.push(cleanField(current));
-            current = '';
-        } else {
-            current += char;
-        }
-        
-        i++;
-    }
-    
-    // 最後のフィールドを追加
-    result.push(cleanField(current));
-    
-    return result;
-}
-
-function cleanField(field) {
-    // 前後の空白を除去
-    field = field.trim();
-    
-    // 前後の引用符を除去（必要に応じて）
-    if (field.startsWith('"') && field.endsWith('"') && field.length >= 2) {
-        field = field.slice(1, -1);
-        // エスケープされた引用符を元に戻す
-        field = field.replace(/""/g, '"');
-    }
-    
-    return field;
-}
 
 function showCSVPreview() {
     const preview = document.getElementById('csvPreview');
@@ -1901,64 +1826,6 @@ function importCSVTasks() {
     }
 }
 
-function isValidDate(dateString) {
-    // 空の場合は有効とする
-    if (!dateString || dateString.trim() === '') {
-        return true;
-    }
-    
-    // スラッシュをハイフンに変換して正規化
-    let normalizedDate = dateString.replace(/\//g, '-');
-    
-    // YYYY-M-D や YYYY-MM-D のような形式を YYYY-MM-DD に変換
-    const parts = normalizedDate.split('-');
-    if (parts.length === 3) {
-        const year = parts[0].padStart(4, '0');
-        const month = parts[1].padStart(2, '0');
-        const day = parts[2].padStart(2, '0');
-        normalizedDate = `${year}-${month}-${day}`;
-    }
-    
-    const regex = /^\d{4}-\d{2}-\d{2}$/;
-    if (!regex.test(normalizedDate)) return false;
-    
-    const date = new Date(normalizedDate + 'T00:00:00');
-    const dateParts = normalizedDate.split('-');
-    const year = parseInt(dateParts[0]);
-    const month = parseInt(dateParts[1]);
-    const day = parseInt(dateParts[2]);
-    
-    // 年、月、日の範囲チェック
-    if (year < 1900 || year > 2100) return false;
-    if (month < 1 || month > 12) return false;
-    if (day < 1 || day > 31) return false;
-    
-    // 実際の日付として有効かチェック
-    return date.getFullYear() === year && 
-           date.getMonth() === month - 1 && 
-           date.getDate() === day;
-}
-
-function normalizeDateString(dateString) {
-    // 空の場合はそのまま返す
-    if (!dateString || dateString.trim() === '') {
-        return dateString;
-    }
-    
-    // スラッシュをハイフンに変換
-    let normalized = dateString.replace(/\//g, '-');
-    
-    // YYYY-M-D や YYYY-MM-D のような形式を YYYY-MM-DD に変換
-    const parts = normalized.split('-');
-    if (parts.length === 3) {
-        const year = parts[0].padStart(4, '0');
-        const month = parts[1].padStart(2, '0');
-        const day = parts[2].padStart(2, '0');
-        normalized = `${year}-${month}-${day}`;
-    }
-    
-    return normalized;
-}
 
 function resetCSVImport() {
     csvData = [];
